% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeMinimalCovering}
\alias{computeMinimalCovering}
\title{Cover a proportion of a given binary set with the smallest number of features}
\usage{
computeMinimalCovering(mat, alpha = 0.05, maxsol = 100, J = 1)
}
\arguments{
\item{mat}{A binary data matrix with each column corresponding to a sample and each row 
corresponding to a feature.}

\item{alpha}{A value in the [0, 1) range indicating what proportion of samples to be 
considered as outlier. By default alpha = 0.05, indicating 95% of samples to be covered.}

\item{maxsol}{The number of optimal solutions to be returned. Default is 100.}

\item{J}{The number of times each sample is to be covered. By default J=1, indicating
that each sample is to be covered with at least one feature.}
}
\value{
A list.
}
\description{
Function for computing the minimal covering for a given binary data matrix given 
a minimum proportion of samples to cover
}
\examples{

}
